refactor: simplify naming for section types
