+++ src/schema/v2/homeView/sectionTypes/Activity.ts
