+++ src/schema/v2/homeView/sectionTypes/Articles.ts
