+++ src/schema/v2/homeView/sectionTypes/Artists.ts
