+++ src/schema/v2/homeView/sectionTypes/Artworks.ts
