+++ src/schema/v2/homeView/sectionTypes/AuctionResults.ts
