+++ src/schema/v2/homeView/sectionTypes/Cards.ts
