+++ src/schema/v2/homeView/sectionTypes/DiscoverMarketingCollection.ts
