+++ src/schema/v2/homeView/sectionTypes/Galleries.ts
