+++ src/schema/v2/homeView/sectionTypes/MarketingCollections.ts
