+++ src/schema/v2/homeView/sectionTypes/Sales.ts
