+++ src/schema/v2/homeView/sectionTypes/ViewingRooms.ts
