+++ src/schema/v2/homeView/sectionTypes/index.ts
@@ -1,20 +1,20 @@
 import { GraphQLObjectType } from "graphql"
 import { ResolverContext } from "types/graphql"
-import { HomeViewActivitySectionType } from "./HomeViewActivitySectionType"
-import { HomeViewArticlesSectionType } from "./HomeViewArticlesSectionType"
-import { HomeViewArtistsSectionType } from "./HomeViewArtistsSectionType"
-import { HomeViewArtworksSectionType } from "./HomeViewArtworksSectionType"
-import { HomeViewAuctionResultsSectionType } from "./HomeViewAuctionResultsSectionType"
-import { HomeViewCardsSectionType } from "./HomeViewCardsSectionType"
-import { HomeViewDiscoverMarketingCollectionType } from "./HomeViewDiscoverMarketingCollectionType"
-import { HomeViewExploreBySectionType } from "./HomeViewExploreBySectionType"
-import { HomeViewFairsSectionType } from "./HomeViewFairsSectionType"
-import { HomeViewGalleriesSectionType } from "./HomeViewGalleriesSectionType"
-import { HomeViewHeroUnitsSectionType } from "./HomeViewHeroUnitsSectionType"
-import { HomeViewMarketingCollectionsSectionType } from "./HomeViewMarketingCollectionsSectionType"
-import { HomeViewSalesSectionType } from "./HomeViewSalesSectionType"
-import { HomeViewShowsSectionType } from "./HomeViewShowsSectionType"
-import { HomeViewViewingRoomsSectionType } from "./HomeViewViewingRoomsSectionType"
+import { HomeViewActivitySectionType } from "./Activity"
+import { HomeViewArticlesSectionType } from "./Articles"
+import { HomeViewArtistsSectionType } from "./Artists"
+import { HomeViewArtworksSectionType } from "./Artworks"
+import { HomeViewAuctionResultsSectionType } from "./AuctionResults"
+import { HomeViewCardsSectionType } from "./Cards"
+import { HomeViewDiscoverMarketingCollectionType } from "./DiscoverMarketingCollection"
+import { HomeViewExploreBySectionType } from "./ExploreBy"
+import { HomeViewFairsSectionType } from "./Fairs"
+import { HomeViewGalleriesSectionType } from "./Galleries"
+import { HomeViewHeroUnitsSectionType } from "./HeroUnits"
+import { HomeViewMarketingCollectionsSectionType } from "./MarketingCollections"
+import { HomeViewSalesSectionType } from "./Sales"
+import { HomeViewShowsSectionType } from "./Shows"
+import { HomeViewViewingRoomsSectionType } from "./ViewingRooms"
 
 export const homeViewSectionTypes: GraphQLObjectType<any, ResolverContext>[] = [
   HomeViewActivitySectionType,
